+++ lib/pio/open_flow10/set_tos.rb
@@ -4,16 +4,16 @@ require 'pio/open_flow/action'
 module Pio
   module OpenFlow10
     # An action to modify the IP ToS/DSCP field of a packet.
-    class SetIpTos < OpenFlow::Action
+    class SetTos < OpenFlow::Action
       action_header action_type: 8, action_length: 8
       uint8 :type_of_service
       string :padding, length: 3
       hide :padding
 
       def initialize(type_of_service)
-        # ip_tos (IP ToS) value consists of 8 bits, of which only the 6
-        # high-order bits belong to DSCP, the 2 low-order bits must be
-        # zero.
+        # ip_tos (IP ToS) value consists of 8 bits, of which only the
+        # 6 high-order bits belong to DSCP, the 2 low-order bits must
+        # be zero.
         unless type_of_service.unsigned_8bit? && type_of_service % 4 == 0
           fail ArgumentError, 'Invalid type_of_service (ToS) value.'
         end