+++ spec/pio/open_flow10/packet_out_spec.rb
@@ -257,13 +257,13 @@ describe Pio::OpenFlow10::PacketOut do
       Then { result.actions[0].ip_address == '1.2.3.4' }
     end
 
-    context 'with a SetIpTos action' do
+    context 'with a SetTos action' do
       When(:user_options) do
         {
           transaction_id: 0x16,
           buffer_id: 0xffffffff,
           in_port: 0xffff,
-          actions: Pio::OpenFlow10::SetIpTos.new(32),
+          actions: Pio::OpenFlow10::SetTos.new(32),
           raw_data: data_dump
         }
       end
@@ -271,7 +271,7 @@ describe Pio::OpenFlow10::PacketOut do
       Then { result.message_length == 0x58 }
       Then { result.actions_len == 0x8 }
       Then { result.actions.length == 1 }
-      Then { result.actions[0].is_a? Pio::OpenFlow10::SetIpTos }
+      Then { result.actions[0].is_a? Pio::OpenFlow10::SetTos }
       Then { result.actions[0].type_of_service == 32 }
     end